requirements.txt
Django
djangorestframework
mysqlclient
pandas
numpy
scikit-learn
python-dotenv
requests
django-filter